blocage

+++ src/Controller/FileController.php
@@ -11,37 +11,47 @@ use App\Document\File;
 use App\Form\FileType;
 use DateTime;
 use DateTimeZone;
+use Symfony\Component\Security\Core\Security;
 
 #[Route('/file')]
 class FileController extends AbstractController
 {
     #[Route('/', name: 'app_file')]
-    public function index(Request $request, FileRepository $fileRepository)
+    public function index(Request $request, FileRepository $fileRepository, Security $security)
     {
         $file = new File();
-        $form = $this->createForm(FileType::class, $file);
-    
-        $form->handleRequest($request);
-    
-        if ($form->isSubmitted()) {
-            if ($form->isValid()) {
-                // Récupérer le contenu du formulaire
-                $content = $file->getContent();
-    
-                // Enregistrer le contenu dans l'objet File
-                $file->setContent($content);
-    
-                $timezone = new DateTimeZone('Europe/Paris');
-                $now = new DateTime('now', $timezone);
-                $file->setCreationDate($now); // Affecte la date de création actuelle
-                $file->setLastUpdate($now); // Affecte la date de modification actuelle
-    
-                $fileRepository->saveFile($file);
-    
-                return $this->redirectToRoute('app_file_list');
-            }
+        $user = $security->getUser();
+
+        if (!$user) {
+            // Gérer le cas où l'utilisateur n'est pas authentifié
         }
-    
+
+        $file->setUserId($user->getId()); // Définir l'id de l'utilisateur sur le fichier
+
+        // Crée le formulaire de création de fichier
+        $form = $this->createForm(FileType::class, $file);
+
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            // Récupérer le contenu du formulaire
+            $content = $file->getContent();
+
+            // Enregistrer le contenu dans l'objet File
+            $file->setContent($content);
+
+            // Définir la date de création et de dernière modification actuelle
+            $timezone = new DateTimeZone('Europe/Paris');
+            $now = new DateTime('now', $timezone);
+            $file->setCreationDate($now);
+            $file->setLastUpdate($now);
+
+            // Sauvegarder le fichier dans la base de données
+            $fileRepository->saveFile($file);
+
+            return $this->redirectToRoute('app_file_list');
+        }
+
         return $this->render('file/index.html.twig', [
             'controller_name' => 'FileController',
             'form' => $form->createView(),
@@ -49,39 +59,56 @@ class FileController extends AbstractController
     }
 
     #[Route('/list', name: 'app_file_list')]
-    public function getListFiles(FileRepository $fileRepository): Response
+    public function getListFiles(FileRepository $fileRepository, Security $security): Response
     {
-        $listFiles = $fileRepository->findBy([], ['lastUpdate' => 'DESC']); // Trier les fichiers par date de modification décroissante;
+        $user = $security->getUser();
+
+        if (!$user) {
+            // Gérer le cas où l'utilisateur n'est pas authentifié
+        }
+
+        $listFiles = $fileRepository->findBy(['userId' => $user->getId()], ['lastUpdate' => 'DESC']);
+
         return $this->render('file/list.html.twig', [
             'controller_name' => 'Liste des fichiers',
             'listFiles' => $listFiles,
-
         ]);
     }
     #[Route('/update/{id}', name: 'app_file_update', methods: ['POST', 'GET'])]
-    public function update(Request $request, string $id, FileRepository $fileRepository): Response
+    public function update(Request $request, string $id, FileRepository $fileRepository, Security $security): Response
     {
         $file = $fileRepository->find($id);
-    
+
         if (!$file) {
             throw $this->createNotFoundException('Le fichier n\'existe pas');
         }
-    
+
+        $user = $security->getUser();
+
+        if (!$user) {
+            // Gérer le cas où l'utilisateur n'est pas authentifié
+        }
+
+        if ($user->getId() !== $file->getUserId()) {
+            // L'utilisateur actuel n'est pas autorisé à mettre à jour ce fichier
+            throw $this->createAccessDeniedException('Vous n\'êtes pas autorisé à modifier ce fichier');
+        }
+
         $form = $this->createForm(FileType::class, $file);
         $form->handleRequest($request);
-    
+
         if ($form->isSubmitted() && $form->isValid()) {
-            if ($request->query->get('update') === 'true') {
-                $timezone = new DateTimeZone('Europe/Paris');
-                $now = new DateTime('now', $timezone);
-                $file->setLastUpdate($now);
-            }
-    
+            // Mettre à jour la date de dernière modification
+            $timezone = new DateTimeZone('Europe/Paris');
+            $now = new DateTime('now', $timezone);
+            $file->setLastUpdate($now);
+
+            // Sauvegarder les modifications dans la base de données
             $fileRepository->saveFile($file);
-    
+
             return $this->redirectToRoute('app_file_list');
         }
-    
+
         return $this->render('file/update.html.twig', [
             'controller_name' => 'Liste des fichiers',
             'form' => $form->createView(),
@@ -89,12 +116,23 @@ class FileController extends AbstractController
     }
 
     #[Route('/delete/{id}', name: 'app_file_delete')]
-    public function delete(string $id, FileRepository $fileRepository): Response
+    public function delete(string $id, FileRepository $fileRepository, Security $security): Response
     {
         $file = $fileRepository->find($id);
 
         if (!$file) {
             throw $this->createNotFoundException('Le fichier n\'existe pas');
+        }
+
+        $user = $security->getUser();
+
+        if (!$user) {
+            // Gérer le cas où l'utilisateur n'est pas authentifié
+        }
+
+        if ($user->getId() !== $file->getUserId()) {
+            // L'utilisateur actuel n'est pas autorisé à supprimer ce fichier
+            throw $this->createAccessDeniedException('Vous n\'êtes pas autorisé à supprimer ce fichier');
         }
 
         $fileRepository->remove($file);
